Big refactor of CNN components. Tweaks to Localizer.py to make things run more smoothly

+++ src/match_seeker/scripts/OlinWorldMap.py
@@ -625,7 +625,7 @@ class WorldMap(object):
                 continue
             parts = line.split()
             cellNum = parts[0]
-            locList = [int(v) for v in parts[1:]]
+            locList = [float(v) for v in parts[1:]]
             # print("Cell " + cellNum + ": ", locList)
             cellDict[cellNum] = locList
         self.cellData = cellDict
